requirements.txt






##tjs2
streamlit>=1.36.0
sentence-transformers
faiss-cpu
PyPDF2>=3.0.0
numpy>=1.24.0
scikit-learn>=1.2.0
tqdm>=4.65.0
torch
torchvision
torchaudio
groq>=0.5.0
## mbc
datetime
werkzeug>=2.0.0
PyJWT>=2.8.0
pandas





